Create scoring.rb in lib\
