+++ app/models/response_map.rb
@@ -1,5 +1,5 @@
 class ResponseMap < ActiveRecord::Base
-  include ActionView::Helpers::AssignmentHelper
+  include Scoring
 
   has_many :response, foreign_key: 'map_id', dependent: :destroy, inverse_of: false
   belongs_to :reviewer, class_name: 'Participant', foreign_key: 'reviewer_id', inverse_of: false
