+++ spec/lib/scoring_spec.rb
@@ -1,0 +1,4 @@
+describe Scoring do
+
+
+end
